+++ src/aten/Dropout.cpp
@@ -3,7 +3,7 @@
 #include <ATen/core/Tensor.h>
 #include <ATen/core/op_registration/adaption.h>
 #include <ATen/native/TensorIterator.h>
-#include <aten/sycl/Dropout.h>
+#include <aten/sycl/DropoutKernels.h>
 
 namespace at {
 
@@ -11,7 +11,7 @@ namespace at {
     const Tensor& input,
     double p,
     ::std::optional<bool> train) {
-  return at::native::xpu::native_dropout(input, p, train);
+  return at::native::xpu::native_dropout_kernel(input, p, train);
 }
 
 Tensor XPUNativeFunctions::native_dropout_backward(
@@ -26,7 +26,8 @@ Tensor XPUNativeFunctions::native_dropout_backward(
       "grad_output");
   c10::impl::check_and_update_common_device(
       common_device, mask, "xpu::native_dropout_backward", "mask");
-  return at::native::xpu::native_dropout_backward(grad_output, mask, scale);
+  return at::native::xpu::native_dropout_backward_kernel(
+      grad_output, mask, scale);
 }
 
 } // namespace at
