+++ src/aten/sycl/AdaptiveAveragePooling2dKernel.h
@@ -4,7 +4,7 @@
 
 namespace at::native::xpu {
 
-void adaptive_avg_pool2d_backward_out_template(
+void adaptive_avg_pool2d_backward_out_template_kernel(
     Tensor& gradInput,
     const Tensor& gradOutput,
     const Tensor& input);
