+++ src/aten/sycl/DropoutKernels.h
@@ -6,12 +6,12 @@ namespace at {
 namespace native {
 namespace xpu {
 
-std::tuple<Tensor, Tensor> native_dropout(
+std::tuple<Tensor, Tensor> native_dropout_kernel(
     const Tensor& self,
     double p,
     c10::optional<bool> train);
 
-Tensor native_dropout_backward(
+Tensor native_dropout_backward_kernel(
     const Tensor& grad,
     const Tensor& mask,
     double scale);
